refactor, adjust to last review
