+++ console-consumer/src/main/java/io/github/jeqo/poc/tracing/consumer/ConsoleConsumer.java
@@ -5,11 +5,9 @@ import brave.Tracing;
 import brave.kafka.clients.KafkaTracing;
 import com.typesafe.config.Config;
 import java.time.Duration;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Properties;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.stream.Stream;
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
@@ -18,11 +16,8 @@ import org.apache.kafka.common.serialization.StringDeserializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static java.util.stream.Collectors.toList;
-
 public class ConsoleConsumer implements Runnable {
   static final Logger LOG = LoggerFactory.getLogger(ConsoleConsumer.class);
-
   final Properties config;
   final Collection<String> topics;
   final KafkaTracing kafkaTracing;
@@ -31,49 +26,56 @@ public class ConsoleConsumer implements Runnable {
 
   public ConsoleConsumer(Tracing tracing, Config consumerConfig) {
     config = new Properties();
-    config.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG,
-        consumerConfig.getString("bootstrap-servers"));
+    config.put(
+        ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, consumerConfig.getString("bootstrap-servers"));
     config.put(ConsumerConfig.GROUP_ID_CONFIG, consumerConfig.getString("group-id"));
     config.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
     config.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
 
+    // specific case: will commit explicitly after record is processed
+    config.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, false);
+
     tracer = tracing.tracer();
     kafkaTracing = KafkaTracing.newBuilder(tracing).writeB3SingleFormat(true).build();
+
     topics = consumerConfig.getStringList("topics");
   }
 
-  private void printRecord(ConsumerRecord<String, String> record) {
-    var span = kafkaTracing.nextSpan(record).name("print").start();
-    try (var ws = tracer.withSpanInScope(span)) {
-      var ts = Stream.of(record.headers().toArray())
-          .map(h -> String.format("%s:%s", h.key(), Arrays.toString(h.value())))
-          .collect(toList());
-      String headers = String.join(",", ts);
-      System.out.printf("record %s-%s-%s: %s=%s (headers: %s)%n", record.topic(),
-          record.partition(), record.offset(), record.key(), record.value(), headers);
-    } catch (RuntimeException | Error e) {
-      span.error(e);
-      throw e;
-    } finally {
-      span.finish();
-    }
-  }
-
-  @Override public void run() {
+  @Override
+  public void run() {
     LOG.info("Starting Console consumer");
-    try (Consumer<String, String> tracingConsumer =
-             kafkaTracing.consumer(new KafkaConsumer<>(config))) {
-      tracingConsumer.subscribe(topics);
+    Consumer<String, String> tracingConsumer = kafkaTracing.consumer(new KafkaConsumer<>(config));
+    var consumerRebalanceListener = new CustomRebalanceListener((KafkaConsumer) tracingConsumer);
+    try {
+      tracingConsumer.subscribe(topics, consumerRebalanceListener);
       while (running.get()) {
+        // clear state of processed records for each poll
+        consumerRebalanceListener.clearProcessedRecordsOffsets();
         var records = tracingConsumer.poll(Duration.ofSeconds(1));
-        records.forEach(this::printRecord);
+        for (ConsumerRecord<String, String> record : records) {
+          processRecord(record, consumerRebalanceListener);
+        }
+        // if async, can be race condition between consumerRebalanceListener.clearProcessedRecordsOffsets();
+        // todo: investigate possibility to use commitAsync()
+        tracingConsumer.commitSync();
       }
     } catch (RuntimeException | Error e) {
       LOG.warn("Unexpected error in polling loop spans", e);
       throw e;
     } finally {
-      LOG.info("Kafka consumer polling loop stopped. Kafka consumer closed.");
+      LOG.info("Kafka consumer polling loop stopped. Kafka consumer trying commit last processed records");
+      try {
+        consumerRebalanceListener.commitProcessedRecords();
+      } finally{
+        LOG.info("Closing consumer");
+        tracingConsumer.close();
+      }
     }
+  }
+
+  private void processRecord(ConsumerRecord<String, String> record, CustomRebalanceListener customRebalanceListener) {
+    // can be long-time processing
+    customRebalanceListener.addOffset(record.topic(), record.partition(), record.offset());
   }
 
   void stop() {
